requirements.txt
# --- Core Framework ---
langchain
langchain-core
langchain-community
langchain-text-splitters 


# --- LLM API Integrations ---
langchain-groq         # For Groq API (fast generation)
langchain-google-genai # For Gemini API (smart routing)
langchain-ollama       # For Ollama (local fallback)
langchain-huggingface
langchain-cohere # for cohere reranker
cohere #cohere API calling
rank_bm25  #BM25 retriever

# --- Server ---
fastapi                # The API framework
uvicorn[standard]      # The server to run FastAPI

# --- Vector Store & Embeddings ---
langchain_chroma
chromadb      # Local vector database
sentence-transformers  # For running embedding models (like BGE)

# --- Data Loading ---
pypdf                  # To load PDF documents

# --- RAG Evaluation ---
ragas                  # For Phase 6 evaluation

# --- Utilities ---
python-dotenv          # To load API keys from .env
langchainhub           # To easily pull standard prompts
